+++ covador/autils.py
@@ -1,0 +1,6 @@
+from asyncio import coroutines
+
+
+def mark_coro(fn):
+    fn._is_coroutine = getattr(coroutines, '_is_coroutine', True)
+    return fn